feat(backend): Implement EventRepository for database operations
- Create EventRepository class in infrastructure layer with:
  - save() method to persist single event
  - saveBatch() method for bulk insert
  - findAll() with pagination and filters
  - findByShopId() method for shop-specific queries
- Update EventsModule to provide and export EventRepository
- Refactor EventsService to use EventRepository for data access

Closes #72
