+++ packages/backend/src/events/events.module.ts
@@ -3,11 +3,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Event } from './entities/event.entity';
 import { EventsController } from './events.controller';
 import { EventsService } from './events.service';
+import { EventRepository } from './infrastructure/event.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Event])],
   controllers: [EventsController],
-  providers: [EventsService],
-  exports: [EventsService],
+  providers: [EventsService, EventRepository],
+  exports: [EventsService, EventRepository],
 })
 export class EventsModule {}
